LayoutTests/imported/w3c/web-platform-tests/html/cross-origin-embedder-policy/javascript.https-expected.txt

CONSOLE MESSAGE: Cancelled load to https://not-web-platform.test:9443/common/blank.html because it violates the resource's Cross-Origin-Resource-Policy response header.
CONSOLE MESSAGE: Fetch API cannot load https://not-web-platform.test:9443/common/blank.html due to access control checks.

PASS Cross-Origin-Embedder-Policy and javascript: URLs
